backend/requirements.txt
# API和基础依赖
google-generativeai==0.3.2
python-dotenv==1.0.1
tqdm==4.66.2

# 数据处理
numpy==1.24.3
pandas==2.0.3

# NLP工具
jieba==0.42.1
spacy==3.7.4

# 深度学习
torch==2.2.1
transformers==4.37.2

# Web框架
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# 工具库
pathlib==1.0.1

# 注意：需要单独安装 spaCy 中文模型
# python -m spacy download zh_core_web_sm 